Fix canvas benchmark & disable _prefer_js
Canvas ops are clearly faster more often than not.  The only case I
was able to find in which they weren't faster was on an HP/AMD RHEL 6
laptop with Chrome 47.  In that case, Chrome was disabling H/W canvas
acceleration because of some issue with the AMD drivers, so canvas and
JS ops were within about 10 ms of each other.  On all other platforms,
canvas ops are many times faster.